Feature: locks
Based on a FIFO queue by Maged M. Michael and Michael L. Scott which
relies on Double Word CAS:
<http://www.cs.rochester.edu/u/scott/papers/1996_PODC_queues.pdf>

Implements a lock with owner and waitlist with direct control of
fiber suspend/resume states.

Also fixes an issue in `scheduler_reschedule` which could prevent the
current fiber's context to be saved, thus be resumeable by another
thread, which would block the thread and the fiber until the
scheduler could steal some work (which could wait forever).

+++ src/fiber.h
@@ -1,15 +1,19 @@
 #ifndef FIBER_H
 #define FIBER_H
 
+#include <signal.h>
+
 typedef struct fiber fiber_t;
 typedef void (*fiber_main_t)();
 typedef void (*fiber_exit_t)(fiber_t *);
 typedef void (*fiber_run_t)(fiber_t *);
 
 typedef struct fiber {
-    int resumeable;
-    void *stack_top;
+    long resumeable; // don't move: required by context asm
+    void *stack_top; // don't move: required by context asm
+
     stack_t stack;
+
     fiber_main_t proc;
     fiber_exit_t link;
 } fiber_t;
